Add Wehrmacht opfor preset.
This may need some work. Specifically, there are no helicopters or
paratroopers available, and the mission may have some trouble with
that. Even if we add a Fallshirmjäger class (one with a parachute), we
don't have a Ju 52 for them to jump out of.

Aside from that, we may want to add a light AT class that carries a
Panzerfaust rather than a Panzerschreck.