Start with the FOB container in the Huron.
Adds an option to start the campaign with the FOB container loaded in
the Huron (or whatever flying mobile respawn) as ace cargo. Useful if
sling loading is not an option, such as when your mobile respawn is a
plane and not a helicopter.

On by default for our WW2 configuration since the air mobile respawn
is a C47.